+++ server/src/weaverbird/pipeline/pipeline.py
@@ -194,6 +194,8 @@ PipelineStepWithVariables = Annotated[
 
 VOID_REPR = "__VOID__"
 EXCLUDE_CLEANING_FOR = (
+    "$ne",  # for isnotnull (None -> null for mongo)
+    "$eq",  # for isnull (None -> null for mongo)
     "pipeline",
     "localField",
     "foreignField",