+++ server/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog (weaverbird python package)
 
-### Changed
+## [0.33.1] - 2023-07-07
+
+### Fixed
 
 - Fix[Mongo]: Add case insensitive on `matches/notmatches` mongo filter step.
 - Fix[Mongo]: ignore `$ne` and `$eq` when removing `__VOID__` values from a mongo pipeline to keep mongo filter steps for `isnull` and `isnotnull`.